# crm/swagger.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:5000/api/v1/",
      "description": "Devlopment server (uses test data)",
      "variables": {}
    },
    {
      "url": "http://13.126.233.224/api/v1/",
      "description": "Production server (uses live data)",
      "variables": {},
      "basePath": ""
    }
  ],
  "info": {
    "version": "v1",
    "title": "CRM API",
    "description": "This api was developed  by eng : Ahmed Ali  ✌️✌️ ,for protoqit company, @copyright protoqit 2021",
    "termsOfService": "",
    "contact": {},
    "license": { "name": "" }
  },
...
  "paths": {
...
          "/admin": {
            "get": {
              "summary": "getAllAdmin",
              "operationId": "getAllAdmin",
              "parameters": [
                {
                  "name": "pageNumber",
                  "in": "query",
                  "required": true,
                  "style": "form",
                  "schema": { "type": "number" }
                },
                {
                  "name": "pageSize",
                  "in": "query",
                  "required": true,
                  "style": "form",
                  "schema": { "type": "number" }
                }
              ],
              "responses": { "200": { "description": "", "headers": {} } },
              "tags": ["Admin"]
            }
          },
          "/admin/clients": {
            "get": {
              "summary": "get All Client",
              "operationId": "getAllClient",
              "parameters": [
                {
                  "name": "pageNumber",
                  "in": "query",
                  "required": true,
                  "style": "form",
                  "schema": { "type": "number" }
                },
                {
                  "name": "pageSize",
                  "in": "query",
                  "required": true,
                  "style": "form",
                  "schema": { "type": "number" }
                }
              ],
              "responses": { "200": { "description": "", "headers": {} } },
              "tags": ["Admin"]
            }
          },
          "/admin/client/{clientId}": {
            "get": {
              "summary": "Get client profile data",
              "operationId": "Get client profile data",
              "parameters": [
                {
                  "name": "clientId",
                  "in": "path",
                  "required": true,
                  "style": "form",
                  "schema": { "type": "string" }
                }
              ],
              "responses": { "200": { "description": "", "headers": {} } },
             
              "tags": ["Admin"]
            }
          },
          "/admin/client/profile/{clientId}": {
            "patch": {
              "summary": "updateClientProfile",
              "operationId": "updateClientProfile",
              "parameters": [
                {
                  "name": "clientId",
                  "in": "path",
                  "required": true,
                  "style": "form",
                  "schema": { "type": "string" }
                }
              ],
              "responses": { "200": { "description": "", "headers": {} } },
              "requestBody": {
                "required": true,
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "example": { "phoneNumber": "01000414736", "lastName": "aaaa" }
                    },
                    "example": "{\r\n     \"phoneNumber\": \"01000414736\",\r\n    \"lastName\":\"aaaa\"\r\n   \r\n \r\n    \r\n}"
                  }
                }
              },
              "tags": ["Admin"]
            }
          },
...
          "/client/file/{folderId}": {
            "get": {
              "summary": "getAllFile",
              "operationId": "getAllFile",
              "parameters": [
                {
                  "name": "pageNumber",
                  "in": "query",
                  "required": true,
                  "style": "form",
                  "schema": { "type": "number" }
                },
                {
                  "name": "pageSize",
                  "in": "query",
                  "required": true,
                  "style": "form",
                  "schema": { "type": "number" }
                },
                {
                  "name": "folderId",
                  "in": "path",
                  "required": true,
                  "style": "form",
                  "schema": { "type": "string" }
                }
              ],
              "responses": { "200": { "description": "", "headers": {} } },
              "tags": ["File"]
            }
            ,"post": {
              "summary": "upload File",
              "operationId": "uploadFile",
              "parameters": [
                {
                  "name": "folderId",
                  "in": "path",
                  "required": true,
                  "style": "form",
                  "schema": { "type": "string" }
                }
              ],
              "responses": { "200": { "description": "", "headers": {} } },
              "requestBody": {
                "required": true,
                "content": {
                  "multipart/form-data": {
                    "schema": {
                      "type": "object",
                      "required": ["file"],
                      "properties": {
                        "file": { "type": "string", "format": "base64" }
                      }
                    }
                   
                  }
                }
              },
              "tags": ["File"]
            }

          },
...
  },
  "components": {
...
        "securitySchemes": {
          "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-auth-token",
            "scheme": "bearer",
            "description": "Enter JWT token",
            "in": "header"
          }
        }
      },
      "security": [{ "ApiKeyAuth": [] }],
...
}